Created corresponding python file.
Python side interface exposes `detect_primitives` interface
that provides pointers to the outputs of the algorithm. Packed into
`PrimitiveSet` object for convenience. This object should clean
up memory when it is destructed, although it is possible to create
a memory leak still. The memory it consumes can also be manually freed.